// flierly-admin-web/src/features/CrudModule/context/action-types.enum.ts
export enum ActionTypes {
    RESET_STATE = "RESET_STATE",
    UPDATE_FILTERS = "UPDATE_FILTERS",
    RESET_FILTERS = "RESET_FILTERS",
    OPEN_UPDATE_FORM = "OPEN_UPDATE_FORM",
    CLOSE_UPDATE_FORM = "CLOSE_UPDATE_FORM",
    OPEN_ROW_MENU = "OPEN_ROW_MENU",
    CLOSE_ROW_MENU = "CLOSE_ROW_MENU",
    ACTIVATE_BIN_MODE = "ACTIVATE_BIN_MODE",
    DEACTIVATE_BIN_MODE = "DEACTIVATE_BIN_MODE",
    UPDATE_ROW_MENU_CURRENT_RECORD = "UPDATE_ROW_MENU_CURRENT_RECORD",
    UPDATE_ROW_MENU_POSITION = "UPDATE_ROW_MENU_POSITION",
    UPDATE_QUERY_CONDITIONS= "UPDATE_QUERY_CONDITIONS",
  }
  